requirements.txt
absl-py==0.11.0
apache-beam==2.25.0
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
avro-python3==1.10.1
backcall==0.2.0
bleach==3.2.1
cachetools==4.1.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
contextlib2==0.6.0.post1
crcmod==1.7
cycler==0.10.0
Cython==0.29.21
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
dm-tree==0.1.5
docopt==0.6.2
entrypoints==0.3
fastavro==1.2.1
Flask==1.1.2
future==0.18.2
gast==0.3.3
gin-config==0.4.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-cloud-bigquery==2.5.0
google-cloud-core==1.4.4
google-crc32c==1.0.0
google-pasta==0.2.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
h5py==2.10.0
hdfs==2.5.8
httplib2==0.17.4
idna==2.10
importlib-resources==3.3.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
kaggle==1.5.10
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lvis==0.5.3
lxml==4.6.2
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mock==2.0.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.19.4
oauth2client==4.1.3
oauthlib==3.1.0
object-detection==0.1
opencv-python==4.4.0.46
opencv-python-headless==4.4.0.46
opt-einsum==3.3.0
packaging==20.7
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pbr==5.5.1
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.8
proto-plus==1.11.0
protobuf==3.14.0
psutil==5.7.3
py-cpuinfo==7.0.0
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pydot==1.4.1
Pygments==2.7.2
pymongo==3.11.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.4
pywin32==300
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.4
Send2Trash==1.5.0
sentencepiece==0.1.94
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-addons==0.11.2
tensorflow-datasets==4.1.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.10.0
tensorflow-metadata==0.25.0
tensorflow-model-optimization==0.5.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
text-unidecode==1.3
tf-models-official==2.3.0
tf-slim==1.1.0
tornado==6.1
tqdm==4.54.1
traitlets==5.0.5
typeguard==2.10.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
